refactor(TvRelativeTime): remove CSS runtime injection and migrate to extracted CSS build for SSR
