+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `@todovue/tv-relative-time` will be documented in this fi
 This project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
+## [1.1.2] - 2025-11-15
+### 🛠️ Changed
+- Removed CSS injection via `vite-plugin-css-injected-by-js` from the library build. Styles are now provided as a separate CSS file generated by Vite (`dist/*.css`) and must be explicitly imported in the host application (SPA or Nuxt).
+- These changes improve compatibility with SSR/SSG environments, especially Nuxt 3/4, by avoiding runtime CSS injection.
+
+---
 ## [1.1.1] - 2025-10-18
 ### 🛠️ Changed
 - The library build now uses `src/entry.ts` (exports both the component and the plugin) instead of directly exporting the `.vue` file.
@@ -16,8 +22,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Explicit export of the style file: `import '@todovue/tv-relative-time/style.css'`.
 - Documentation for usage in SSR and Nuxt 3 applications.
 
+---
 ## [1.1.0] - 2025-05-06
-
 ### ✨ Features
 - Added localization support for relative time formatting (`es`, `fr`, `pt`, and `en` as default)
 
@@ -26,8 +32,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Watched `date` and other props to ensure reactivity
 - Updated package references and improved documentation consistency
 
+---
 ## [1.0.0] - 2025-05-05
-
 ### ✨ Features
 - Fully functional `<TvRelativeTime />` component.
 - Live update support with customizable `updateInterval`.
@@ -44,6 +50,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
 
+[1.1.2]: https://github.com/TODOvue/tv-relative-time/pull/4/files
 [1.1.1]: https://github.com/TODOvue/tv-relative-time/pull/3/files
 [1.1.0]: https://github.com/TODOvue/tv-relative-time/pull/2/files
 [1.0.0]: https://github.com/TODOvue/tv-relative-time/pull/1/files
