--- src/assets/scss/_variables.scss
@@ -1,3 +1,0 @@
-/* 🖋 Fonts */
-$font-text: 'Lato', sans-serif;
-$font-title: 'Kanit', sans-serif;
