--- src/assets/scss/styles.scss
@@ -1,6 +1,0 @@
-@use "variables";
-
-.tv-relative-time {
-  cursor: help;
-  font-family: variables.$font-text;
-}
