+++ src/demo/Demo.vue
@@ -1,8 +1,7 @@
 <script setup>
-import { defineAsyncComponent } from "vue";
-import { TvDemo } from '@todovue/tv-demo';
-const TvRelativeTime = defineAsyncComponent(() => import("../components/TvRelativeTime.vue"));
-import { demos } from '../utils/mocks.js';
+import { defineAsyncComponent } from 'vue';
+const TvRelativeTime = defineAsyncComponent(() => import('../components/TvRelativeTime.vue'));
+import { demos } from './utils/mocks.js';
 </script>
 
 <template>
@@ -13,7 +12,7 @@ import { demos } from '../utils/mocks.js';
     npm-install="@todovue/tv-relative-time"
     source-link="https://github.com/TODOvue/tv-relative-time"
     url-clone="https://github.com/TODOvue/tv-relative-time.git"
-    version="1.1.1"
+    version="1.1.2"
   />
 </template>
