+++ src/demo/utils/demos/compact.vue
