+++ src/demo/utils/demos/french.vue
