+++ vite.config.js
@@ -1,18 +1,17 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
-import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
 import dts from "vite-plugin-dts";
 const isDemo = process.env.VITE_BUILD_TARGET === "demo";
 
 export default defineConfig({
   plugins: [
     vue(),
-    cssInjectedByJsPlugin(),
     dts({
       insertTypesEntry: true,
       outputDir: "dist",
       skipDiagnostics: false,
-    })],
+    })
+  ],
   build: isDemo
     ? {
       outDir: "dist-demo",
